Added backspace support

+++ cmfOS/kernel/src/std/math.h
@@ -2,8 +2,8 @@
 #include "ctypedef.h"
 
 struct Point {
-    unsigned int x; 
-    unsigned int y; 
+    unsigned long x; 
+    unsigned long y; 
 };
 
 double power(double base, int exponent);
